Add get_usergroup_of_users
